+++ app/src/main/java/name/gudong/translate/GDApplication.java
@@ -29,10 +29,10 @@ import com.orhanobut.logger.Logger;
 import im.fir.sdk.FIR;
 import jonathanfinerty.once.Once;
 import me.gudong.translate.BuildConfig;
-import name.gudong.translate.reject.components.AppComponent;
-import name.gudong.translate.reject.components.DaggerAppComponent;
-import name.gudong.translate.reject.modules.ApiServiceModel;
-import name.gudong.translate.reject.modules.AppModule;
+import name.gudong.translate.injection.components.AppComponent;
+import name.gudong.translate.injection.components.DaggerAppComponent;
+import name.gudong.translate.injection.modules.ApiServiceModel;
+import name.gudong.translate.injection.modules.AppModule;
 
 /**
  * Created by GuDong on 12/27/15 16:46.
