+++ app/src/main/java/name/gudong/translate/injection/modules/AppModule.java
@@ -18,7 +18,7 @@
  *
  */
 
-package name.gudong.translate.reject.modules;
+package name.gudong.translate.injection.modules;
 
 import com.litesuits.orm.LiteOrm;
