+++ app/src/main/java/name/gudong/translate/ui/activitys/AcknowledgementsActivity.java
@@ -24,8 +24,8 @@ import butterknife.ButterKnife;
 import me.gudong.translate.R;
 import name.gudong.translate.mvp.model.entity.Acknowledgement;
 import name.gudong.translate.mvp.presenters.AcknowledgementPresenter;
-import name.gudong.translate.reject.components.AppComponent;
-import name.gudong.translate.reject.modules.ActivityModule;
+import name.gudong.translate.injection.components.AppComponent;
+import name.gudong.translate.injection.modules.ActivityModule;
 import name.gudong.translate.ui.adapter.AcknowledgementAdapter;
 import name.gudong.translate.ui.common.DefaultItemDecoration;
 import rx.Observable;
